Add files via upload
This Python script allows a user to search for images on Google anaonymously by using the Tor network.  
It prompts the user to enter a search keyword, performs the search through Google's image search, scrapes the resulting image URLs and downloads the images to a local folder.

The anonymity is provided by routing all the search and download requests through Tor, esnuring that the user's IP address is masked.

--Tor (The Onion Router)--
This network, anonymizes internet traffic by routing it through a series of volunteer-operated servers(nodes), making it difficult to trace the user's location or online activity.  
Tor was implemented here for privacy protection, allowing users to browse the internet without revealing their identity.