Add files via upload
This Python script is designed to transcribe audio files and translates the transcription using the Whisper model and Google Translate.  
It leverages libraries such as Whisper for automation speech recognition, GoogleTrans for translation, ffmpeg for audio format conversion, and tkinter for a user friendly graphical interface.  

The script begins by loading the Whisper model  with the options to choose different versions by varying levels of accuracy. It inculdes a function to convert audio files to a compatible format using ffmpeg, esnuring that the input is suitable for transcription, The transcription function processes the audio file extracting spoken content and returning it as text in English.

In addition to transcription, the script offers translation capabilities. It takes the transcribed text and translates it into a specified target language, such as Hindi, using the Google translate API. The file upload function allows user to select audio files from their system through a simple dialog box, supporting multiple audio formats like WAV and MP3.

-----***-----
It also handles clean up by removing temporary files created during the process. Overall, this script provides an efficient and straightforward way to convert spoken lanaguage into text and translate it into another language.

---Libraries Used---
Whisper : A Library for automatic speech recognition(ASR) that provides functionality to transcribe audio files.
Googletrans: A Library that aloows you to interface with the Google Translate API for translating text.
fame : A multimedia framework to convert audio formats.
tkinter : A standard Python interface to create a GUI for file selection.